scripts/requirements.txt
html-linter==0.4.0
pytidylib==0.3.2
requests==2.23.0
typed-ast>=1.4.1
urllib3>=1.25.9

